packages/backend/pyproject.toml
[project]
name = "cs2-backend"
version = "0.1.0"
description = "AI CS2 Analytics — Backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.7.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "python-jose[cryptography]>=3.3.0",
    "argon2-cffi>=23.1.0",
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",
    "slowapi>=0.1.9",
    "python-multipart>=0.0.18",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "aiosqlite>=0.21.0",
]
dev = [
    "ruff>=0.11.0",
    "mypy>=1.14.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
